Refs #5 - separated the CiviCRM entities from the API
I moved Contact and Address to Chiro.CiviCrm.Domain, and I share these classes
between Chiro.CiviCrm.ClientInterfaces and Chiro.CiviCrm.Api.

I am not sure whether this is the best solution. I could use different classes for the
API (speaking to CiviCRM) and the client interface (speaking to the user), and
do some mapping. This way a Contact can have a list of Addresses and so on.
But at this moment, I think that sharing the classes is OK.