Closes #49 - Filtering infrastructure is there.
In a request you can change the type T of a field you want to be able to
filter to Filter<T>. I am not sure whether it is a good idea to do this for
all fields. But it is already done for start and end date of events.